Update fsst12.h
windows fix (long is 32-bits -- use long long instead)

+++ fsst12.h
@@ -154,7 +154,7 @@ fsst_decompress(
    unsigned char *output    /* OUT: memory buffer to put the decompressed string in. */
 ) {
    unsigned char*__restrict__ len = (unsigned char* __restrict__) decoder->len;
-   unsigned long*__restrict__ symbol = (unsigned long* __restrict__) decoder->symbol; 
+   unsigned long long*__restrict__ symbol = (unsigned long long* __restrict__) decoder->symbol; 
    unsigned char*__restrict__ strOut = (unsigned char* __restrict__) output;
    unsigned long posOut = 0, posIn = 0;
 #define FSST_UNALIGNED_STORE(dst,src) memcpy((unsigned long long*) (dst), &(src), sizeof(unsigned long long))
